Fix bug in runOneStep by setting time and step first
Semantically, an activity runs at a given time. The schedule
parameter -- given to the activity -- should reflect this time.
Prior to this commit, the time and step were set **after**
execution. Not only is this semantically wrong, but it introduces
schedule bugs via delta-based schedulers. For example, `onceIn`
with a delta of 1 is set relative to the previous time, not now.